Fixed getting location

+++ app/src/main/java/com/example/weatherapp/data/Api.kt
@@ -9,11 +9,12 @@ interface Api {
     // api.openweathermap.org/data/2.5/forecast?lat={lat}&lon={lon}&appid={API key}
     @GET("forecast")
     suspend fun getWeatherForecast(
-        @Query("lat") latitude: Int,
-        @Query("lon") longitude: Int,
+        @Query("lat") latitude: Double,
+        @Query("lon") longitude: Double,
         @Query("appid") apikey: String
     ): WeatherForecast
 
+    // todo
     // api.openweathermap.org/data/2.5/forecast?q={city name}&appid={API key}
     @GET("forecast")
     suspend fun getWeatherForecast(
