+++ app/src/main/java/com/example/weatherapp/data/WeatherRepository.kt
@@ -6,7 +6,7 @@ private const val API_KEY = "b65750fa8912403310944973c7362a56"
 
 class WeatherRepository(private val api: Api) {
 
-    suspend fun getWeatherForecast(latitude: Int, longitude: Int): WeatherForecast {
+    suspend fun getWeatherForecast(latitude: Double, longitude: Double): WeatherForecast {
         return api.getWeatherForecast(latitude, longitude, API_KEY)
     }
